% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{alter_order_PA_vector}
\alias{alter_order_PA_vector}
\title{Alter order of compositional column labels.}
\usage{
alter_order_PA_vector(comp_labels, component_1)
}
\arguments{
\item{comp_labels}{List of compositional column labels.}

\item{component_1}{Component which should be moved to front.}
}
\value{
\code{comp_labels} in new order.
}
\description{
Alter order of compositional column labels.
}
