% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr.R
\name{clr_trans}
\alias{clr_trans}
\title{Clr transformation}
\usage{
clr_trans(data)
}
\arguments{
\item{data}{Compositional columns of dataframe.}
}
\value{
\code{data} with clr-transformed compositional columns.
}
\description{
This performs centred log-ratio transformation on the compositional columns of a data frame.
}
\examples{
clr_trans(data = simdata[, c("vigorous", "moderate", "light", "sedentary", "sleep")])

}
