% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr.R
\name{clr_trans_inv}
\alias{clr_trans_inv}
\title{Clr inversion}
\usage{
clr_trans_inv(data)
}
\arguments{
\item{data}{Clr-transformed columns.}
}
\value{
\code{data} with clr transformation inverted.
}
\description{
This inverts a clr transformation. Currently it only does so onto the unitless scale. If needed, it can be extended to take a units output argument.
}
