% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_mean.R
\name{comp_mean}
\alias{comp_mean}
\title{Compositional mean}
\usage{
comp_mean(data, comp_labels)
}
\arguments{
\item{data}{Dataset to calculate compositional mean of.}

\item{comp_labels}{The labels of the compositional component.}
}
\value{
Vector which is the compositional mean.
}
\description{
Calculates the compositional mean of a dataset.
}
\examples{
#TBA
}
