% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_model.R
\name{comp_model}
\alias{comp_model}
\title{Statistical models with compositional exposure variables}
\usage{
comp_model()
}
\arguments{
\item{composition}{This should be the starting composition as a data frame.}

\item{main_part}{part the main change is happening to.}

\item{main_change}{Magnitude and direction of change e.g. +1, -0.5.}

\item{comp_labels}{Compositional column labels.}
}
\value{
Composition after rebalancing in the described way.
}
\description{
This is a wrapper for \code{lm}, \code{glm} and \code{survival::coxph} which
}
