% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_plotting.R
\name{generate_fixed_values}
\alias{generate_fixed_values}
\title{Generates list of fixed_values based on median/modal values in dataset}
\usage{
generate_fixed_values(
  data,
  comp_labels,
  rounded_zeroes,
  det_limit,
  units,
  specified_units
)
}
\arguments{
\item{data}{Data used for model development.}

\item{comp_labels}{The labels of the compositional columns.}

\item{rounded_zeroes}{Are zeroes rounded zeroes?}

\item{det_limit}{Detection limit if zeroes are to be imputed. This must be set if \code{rounded_zeroes} is \code{TRUE} and should be the
minimum measurable value in the compositional columns of data.}

\item{units}{What are the units of the compositional variables? Currently available are "unitless" (where working in terms of proportions), "hr/day", "hr/wk", "min/day", "min/wk" and "specified", in which case the \code{specified_units} argument should be set.}

\item{specified_units}{If units are being specified via the composition sum, this is where it is done. It should be a vector where the first argument is a string describing the units, and the second argument is the expected sum of a composition e.g. \code{c("hr/day", 24)}}
}
\value{
dataframe with a single row of fixed_values.
}
\description{
If fixed values are not set, this will set them to modal/ median values.
}
