% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilr.R
\name{ilr_trans_inv}
\alias{ilr_trans_inv}
\title{ilr_trans_inv: Inverts ilr transformations (pivot coordinates)}
\usage{
ilr_trans_inv(data)
}
\arguments{
\item{data}{Dataset to have the transformation reversed}
}
\value{
Compositional columns
}
\description{
Takes ilr-transformed columns and inverts the ilr transformation.
}
\examples{
# TBA
}
