% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_plotting.R
\name{make_new_data}
\alias{make_new_data}
\title{make_new_data: Generates a new dataset varying in the dimension of interest.}
\usage{
make_new_data(from_component, to_component, fixed_values, dataset,
  units = NULL, comp_labels, lower_quantile = 0.05,
  upper_quantile = 0.95, granularity = 10000)
}
\arguments{
\item{from_component}{Component to plot transfer out of. Of the two being considered, should be component with generally higher values (e.g. higher median)}

\item{to_component}{Component to plot transfer into. Of the two being considered, should be component with generally lower values (e.g. lower median.)}

\item{fixed_values}{Should be a dataframe with a column for each component and containing a single value for each of those.}

\item{dataset}{Dataset to take distribution of \code{from_component} from.}

\item{comp_labels}{Labels of compositional columns.}

\item{lower_quantile}{See \code{vary_time_of_interest}}

\item{upper_quantile}{See \code{vary_time_of_interest}}

\item{granularity}{Doesn't usually need setting. Parameter indicating how many predictions to make. If too low, plotted curve has gaps. If too high, calculation is slow.}

\item{comp_sum}{Numeric value indicating the value compositional columns should sum to (e.g. 1, 24, 168, 10080).}
}
\value{
Dataframe varying in time of interest.
}
\description{
Generates a new dataset to feed into the plotting functions.
}
\examples{

}
