% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalise_comp}
\alias{normalise_comp}
\title{Normalise input data}
\usage{
normalise_comp(data, comp_labels)
}
\arguments{
\item{data}{Input data.}

\item{comp_labels}{The labels of the compositional columns.}
}
\description{
Normalise input data
}
