% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_axis_label}
\alias{process_axis_label}
\title{Process axis labels}
\usage{
process_axis_label(label, terms, type)
}
\arguments{
\item{label}{This is the axis label that already exsits.}

\item{terms}{Are predictions for terms,or are they absolute?}

\item{type}{Output from \code{process_model_type}}
}
\description{
Process axis labels
}
