% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_model_type}
\alias{process_model_type}
\title{Process model argument}
\usage{
process_model_type(model)
}
\arguments{
\item{model}{This is the model which needs type extracting.}
}
\description{
Process model argument
}
