% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_units}
\alias{process_units}
\title{Process units argument}
\usage{
process_units(units, specified_units)
}
\arguments{
\item{units}{List of compositional column labels.}

\item{specified_units}{part which should be moved to front.}
}
\description{
Process units argument
}
