% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prompt_scale}
\alias{prompt_scale}
\title{Prompt scale of axes depending on model type}
\usage{
prompt_scale(model, terms, plot_log)
}
\arguments{
\item{model}{Model to use in estimates/predictions.}

\item{terms}{Are estimates for differences in outcome associated with differences in compositional variables? If \code{terms = TRUE} all estimates and plots will be for difference in outcome associated with differences in the compositional variables. If \code{terms = FALSE}, \code{fixed_values} is used to set the values of the non-compositional covariates, and outputs are predictions for the outcome based on these values of the non-compositional covariates and the given value of the compositional variables (and confidence intervals include uncertainty due to all variables in the model, not just the compositional variables). Note that for logistic regression models with \code{terms = TRUE} estimates are odds ratios; for logistic regression models with \code{terms = FALSE} estimates are probabilities (i.e. predictions on the response scale).}

\item{plot_log}{If this is \code{TRUE}, the y-axis will be log-transformed.}
}
\description{
Prompt scale of axes depending on model type
}
