% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rescale_comp}
\alias{rescale_comp}
\title{Return output data on correct scale}
\usage{
rescale_comp(data, comp_labels, comp_sum)
}
\arguments{
\item{data}{Data including a normalised set of compositional columns.}

\item{comp_labels}{The labels of the compositional columns.}

\item{comp_sum}{The sum the compositional columns should have.}
}
\description{
Return output data on correct scale
}
