% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plots.R
\name{save_plot}
\alias{save_plot}
\title{Save plots with desired quality.}
\usage{
save_plot(plot, dpi = 300, filename, filetype = "tiff")
}
\arguments{
\item{plot}{Object output from a plotting function.}

\item{dpi}{Resolution.}

\item{filename}{Desired path to saved file.}

\item{filetype}{Should be "tiff", "png", "pdf", "jpg" etc.}
}
\value{
Saves file with the desired size.
}
\description{
A wrapper for ggsave, to make saving with desired quality easier.
}
\details{
ggsave or a routine such as tiff("filename") , plot, dev.off() can be used as usual to save plots; there is no need to employ this wrapper. However, this is intended to make saving higher quality images easier.
}
