% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_of_data.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{Example compositional data}
\format{
An object of class \code{data.frame} with 10000 rows and 12 columns.
}
\usage{
simdata
}
\description{
This is a simulated physical behaviour dataset.
}
\keyword{data}
