% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_of_data.R
\docType{data}
\name{simdata_ilr}
\alias{simdata_ilr}
\title{Example compositional data with transformed variables}
\format{An object of class \code{data.frame} with 10000 rows and 14 columns.}
\usage{
simdata_ilr
}
\description{
This is a simulated time use dataset, with ilr pivot coordinates included.
}
\keyword{data}
