% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_of_coefficients.R
\name{tab_covariate_coeffs}
\alias{tab_covariate_coeffs}
\title{Tabulate coefficients associated with covariates}
\usage{
tab_covariate_coeffs(
  model = NULL,
  scale_type = "lp",
  comp_labels,
  transformation_type = "ilr",
  rounded_zeroes = TRUE,
  det_limit = NULL,
  comparison_part = NULL,
  part_1 = NULL
)

tab_covariate_coeffs(
  model = NULL,
  scale_type = "lp",
  comp_labels,
  transformation_type = "ilr",
  rounded_zeroes = TRUE,
  det_limit = NULL,
  comparison_part = NULL,
  part_1 = NULL
)
}
\arguments{
\item{scale_type}{Can take value "lp" (linear predictors) or "exp" (exponentiated- relevant for interpretation of logistic and Cox models).}

\item{inheritParams}{comp_model}
}
\value{
Table of covariates and their associated coefficients (parameter value with lower and upper confidence interval of the 95% ).

Table of covariates and their associated coefficients (parameter value with lower and upper confidence interval of the 95% ).
}
\description{
This pulls out a table of the coefficients of the covariates only in the model.

This pulls out a table of the coefficients of the models.
}
