% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_of_coefficients.R
\name{tab_covariate_coefs}
\alias{tab_covariate_coefs}
\title{Tabulate coefficients associated with covariates}
\usage{
tab_covariate_coefs(
  model = NULL,
  scale_type = "lp",
  level = 0.95,
  comp_labels,
  part_1 = NULL
)
}
\arguments{
\item{model}{Model for which extracting coefficients.}

\item{scale_type}{Can take value "lp" (linear predictors) or "exp" (exponentiated- relevant for interpretation of logistic and Cox models).}

\item{level}{The level argument of the confidence intervals. Passed directly to `stats::confint`.}

\item{comp_labels}{The labels of the compositional columns.}

\item{part_1}{Used with ilr-pivot coordinate transformation to specify which should be first part. Passed to \code{alter_order_comp_labels} unless \code{NULL}.}
}
\value{
Table of covariates and their associated coefficients (parameter value with lower and upper confidence interval). Note that, unlike the plotting and some other default model summary functions in R, this uses confidence intervals based on stats::confint.
}
\description{
This pulls out a table of the coefficients of the covariates only in the model.
}
