% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tmat}
\alias{tmat}
\title{Create transformation matrix from clr to ilr pivot coordinates}
\usage{
tmat(cols)
}
\arguments{
\item{nrow}{Number of rows of transformation matrix to be created (= number of parts in composition)}
}
\description{
New improved version of create_transformation_matrix due to Ben Feakins
}
