% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_plotting.R
\name{vary_component_of_interest}
\alias{vary_component_of_interest}
\title{Varies component of interest.}
\usage{
vary_component_of_interest(component_of_interest, lower_quantile = 0.05,
  upper_quantile = 0.95)
}
\arguments{
\item{component_of_interest}{The variable of interest.}

\item{lower_quantile}{The lower quantile to vary from.}

\item{upper_quantile}{The upper quantile to vary to.}
}
\value{
Vector of values going from \code{lower_quantile} to \code{upper_quantile} of the distribution of the varaible of interest.
}
\description{
Produces variable going between percentiles of the component of interest in the data.
}
