% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf_variables.R
\name{vector_to_sum}
\alias{vector_to_sum}
\title{Create sum from a vector}
\usage{
vector_to_sum(vector)
}
\arguments{
\item{vector}{Vector to be turned into a sum.}
}
\value{
Sum of elements of vector.
}
\description{
Create sum from a vector
}
