% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zero_range}
\alias{zero_range}
\title{Determine if range of vector is effectively zero.}
\usage{
zero_range(x, tol = 0.01)
}
\arguments{
\item{x}{Data for which to check}

\item{tol}{Tolerance within which considered equal (on median-scaled scale).}
}
\description{
Determine if range of vector is effectively zero.
}
