% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alr.R
\name{alr_trans}
\alias{alr_trans}
\title{Alr transformation}
\usage{
alr_trans(data, comparison_component = NULL, comp_labels)
}
\arguments{
\item{data}{Compositional columns of dataframe}

\item{comparison_component}{Name of component that all other components will be compared to.}

\item{comp_labels}{The labels of the compositional columns.}
}
\description{
Alr transformation
}
