% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alr.R
\name{alr_trans}
\alias{alr_trans}
\title{Alr transformation}
\usage{
alr_trans(data, comp_labels, comparison_part = NULL)
}
\arguments{
\item{data}{Compositional columns of dataframe}

\item{comp_labels}{The labels of the compositional columns.}

\item{comparison_part}{Name of part that all other parts will be compared to.}
}
\value{
\code{data} with alr-transformed compositional columns.
}
\description{
This peforms additive log-ratio transformation.
}
\examples{
transform_comp(data = simdata,
comp_labels = c("partA", "partB", "partC", "partD", "partE"),
comparison_part = "partA")

}
