% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_composition.R
\name{change_composition}
\alias{change_composition}
\title{Change composition}
\usage{
change_composition(
  composition,
  main_part,
  at_expense_of = NULL,
  main_change,
  comp_labels
)
}
\arguments{
\item{composition}{This should be the starting composition as a data frame.}

\item{main_part}{part the main change is happening to.}

\item{at_expense_of}{Parts the main change is at the expense of/ reallocated to.}

\item{main_change}{Magnitude and direction of change e.g. +1, -0.5.}

\item{comp_labels}{Compositional column labels.}
}
\value{
Composition after reallocating as described.
}
\description{
This makes a change to a composition in the specified way and automatically reallocates to/from other parts of the composition proportionally. It primarily exists as a utility function to make reporting easier.
}
\examples{
new_comp <-
change_composition(
 composition = simdata,
 main_part = "moderate",
 main_change = +0.5,
 comp_labels = c("vigorous", "moderate", "light", "sedentary", "sleep")
 )
}
