% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_composition.R
\name{change_composition}
\alias{change_composition}
\title{Change composition}
\usage{
change_composition(composition, main_part, main_change, comp_labels)
}
\arguments{
\item{composition}{This should be the starting composition as a data frame.}

\item{main_part}{part the main change is happening to.}

\item{main_change}{Magnitude and direction of change e.g. +1, -0.5.}

\item{comp_labels}{Compositional column labels.}
}
\value{
Composition after rebalancing in the described way.
}
\description{
This makes a change to a composition in the specified way and automatically rebalances.
}
