% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_mean.R
\name{comp_mean}
\alias{comp_mean}
\title{Compositional mean}
\usage{
comp_mean(
  data,
  comp_labels,
  rounded_zeroes = TRUE,
  det_limit = NULL,
  units = "unitless",
  specified_units = NULL
)
}
\arguments{
\item{data}{Dataset to calculate compositional mean of.}

\item{comp_labels}{The labels of the compositional parts.}

\item{rounded_zeroes}{Are zeroes rounded zeroes?}

\item{det_limit}{Detection limit if zeroes are to be imputed. This must be set if \code{rounded_zeroes} is \code{TRUE} and should be the
minimum measurable value in the compositional columns of data. It should be on the same scale as the (input) compositional columns.}

\item{units}{What should the units of the compositional variables be in any output? Currently available are "unitless" (where working in terms of proportions), "hr/day", "hr/wk", "min/day", "min/wk" and "specified", in which case the \code{specified_units} argument should be set. Note that this doesn't specify the input units, as this is not relevant for any function.}

\item{specified_units}{If units are being specified via the composition sum, this is where it is done. It should be a vector where the first argument is a string describing the units, and the second argument is the expected sum of a composition e.g. \code{c("hr/day", 24)}}
}
\value{
Vector which is the compositional mean.
}
\description{
Calculates the compositional mean of a dataset.
}
\examples{
comp_mean(data = simdata,
          comp_labels = c("partA", "partB", "partC", "partD", "partE"),
          units = "hr/day")

comp_mean(data = simdata, # this is the dataset
comp_labels = comp_labels, # this is the labels of the compositional columns,
# which we specified above
rounded_zeroes = TRUE, # this option specifies that we'll treat the zeroes
# as rounded zeroes i.e. we'll impute them
det_limit = 0.0083, # this is the smallest value observable in the data
units = "hr/day" # this is the units. There are pre-specified options "hr/day",
# "hr/wk", "min/day", "min/wk" and "unitless".
# If you set units = "specified", you can also specify your own units using
# specified = c("my_units_name", sum of a composition in these units)
)
}
