% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_model.R
\name{comp_model}
\alias{comp_model}
\title{Statistical models with compositional exposure variables}
\usage{
comp_model(
  type = NULL,
  outcome = NULL,
  covariates = NULL,
  follow_up_time = NULL,
  event = NULL,
  data,
  comp_labels,
  transformation_type = "ilr",
  rounded_zeroes = TRUE,
  det_limit = NULL,
  comparison_part = NULL,
  part_1 = NULL
)
}
\arguments{
\item{type}{Type of statistical model to use. It should be one of "linear", "logistic", "cox".}

\item{outcome}{Column name of outcome variable in data. If \code{type} is \code{"linear"}, this should be a continuous variable.  If \code{type} is \code{"logistic"}, this should be a binary outcome.  If \code{type} is \code{"cox"}, if this is set it should be a \code{Surv} object from package \code{survival} (if it is set, the function defaults to attempt to use it, even if \code{follow_up_time} and \code{event} are set). If this is left as \code{NULL}, \code{follow_up_time} and \code{event} can be set instead.}

\item{covariates}{Character vector of column names of covariates to adjust models for. As this is used as a character vector, special arguments to the standard models (like strata(variable) for a Cox model) can be used here.}

\item{follow_up_time}{Only used if \code{type} is \code{"cox"} and \code{outcome} is \code{NULL}.  Follow-up time.}

\item{event}{Only used if \code{type} is \code{"cox"} and \code{outcome} is \code{NULL}.Binary variable indicating whether or not an event was observed.}

\item{data}{Dataset to have the transformation applied to.}

\item{comp_labels}{The labels of the compositional columns.}

\item{transformation_type}{The type of transformation desired. Should be \code{"ilr", "alr"} or \code{"clr"}.}

\item{rounded_zeroes}{Are zeroes rounded zeroes?}

\item{det_limit}{Detection limit if zeroes are to be imputed. This must be set if \code{rounded_zeroes} is \code{TRUE} and should be the
minimum measurable value in the compositional columns of data. It should be on the same scale as the (input) compositional columns.}

\item{comparison_part}{Only needed for alr transformation. Should be an element of \code{comp_labels}. Name of part that all other parts will be compared to.}

\item{part_1}{Only used if First part in ilr-pivot coordinate transformation. Passed to \code{alter_order_comp_labels} unless \code{NULL}.}
}
\value{
model
}
\description{
This is a wrapper for \code{lm}, \code{glm} and \code{survival::coxph} which performs the compositional transformation before generating the model.
}
\examples{
 comp_model(type = "linear",
outcome = "linear_outcome",
covariates = c("agegroup", "sex"),
follow_up_time = NULL,
event = NULL,
data = simdata,
comp_labels = c("partA", "partB", "partC", "partD", "partE"),
rounded_zeroes = TRUE,
det_limit = 0.0083)
}
