% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_transformation_matrix}
\alias{create_transformation_matrix}
\title{Create transformation matrix from clr to ilr pivot coordinates}
\usage{
create_transformation_matrix(nrow)
}
\arguments{
\item{nrow}{Number of rows of transformation matrix to be created (= number of parts in composition)}
}
\description{
Create transformation matrix from clr to ilr pivot coordinates
}
