% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot_examples.R
\name{forest_plot_comp}
\alias{forest_plot_comp}
\title{Produce a forest plot indicating model prediction at given compositions}
\usage{
forest_plot_comp(
  composition_list,
  model = NULL,
  models_list = NULL,
  dataset,
  fixed_values = NULL,
  transformation_type = NULL,
  comparison_part = NULL,
  part_1 = NULL,
  comp_labels,
  units = "unitless",
  specified_units = NULL,
  rounded_zeroes = TRUE,
  det_limit = NULL,
  terms = TRUE,
  x_label = NULL,
  xllimit = NULL,
  xulimit = NULL,
  plot_log = FALSE,
  text_settings = NULL,
  pred_name = NULL,
  ...
)
}
\arguments{
\item{composition_list}{Named list of compositions. Note each composition should be stored as a data frame. For example, use the output of \code{change_composition}.}

\item{model}{Model to use in generating predictions.}

\item{models_list}{If \code{model} is \code{NULL} (or not set), a named list of models for which to plot model predictions in the forest plot. Note all models should have the same type or the results will be meaningless.}

\item{dataset}{Should be dataset used to develop \code{model}.}

\item{fixed_values}{If \code{terms = FALSE}, this is used as giving the fixed values of the non-compositional covariates at which to calculate the prediction. If it is not set, it can be automatically generated.}

\item{transformation_type}{The type of transformation desired. Should be \code{"ilr", "alr"} or \code{"clr"}.}

\item{comparison_part}{Only needed for alr transformation. Should be an element of \code{comp_labels}. Name of part that all other parts will be compared to.}

\item{part_1}{Only used if First part in ilr-pivot coordinate transformation. Passed to \code{alter_order_comp_labels} unless \code{NULL}.}

\item{comp_labels}{The labels of the compositional columns.}

\item{rounded_zeroes}{Are zeroes rounded zeroes?}

\item{det_limit}{Detection limit if zeroes are to be imputed. This must be set if \code{rounded_zeroes} is \code{TRUE} and should be the
minimum measurable value in the compositional columns of data. It should be on the same scale as the (input) compositional columns.}

\item{terms}{Are predictions for terms,or are they absolute?}

\item{x_label}{Label for x axis in plot.}

\item{xllimit}{Minimum value for x axis.}

\item{xulimit}{Maximum value for x axis.}

\item{plot_log}{If this is \code{TRUE}, the x-axis will be log-transformed.}

\item{text_settings}{An optional argument which should be an \code{fpTxtGp} object as specified in the \code{forestplot} package.}

\item{...}{
  Arguments passed on to \code{\link[forestplot:forestplot]{forestplot::forestplot}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
Forest plot illustrating prediction of the model at given compositions.
}
\description{
This function takes a named list of compositions, and plots a model prediction at each composition.
}
