% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_plotting.R
\name{generate_fixed_values}
\alias{generate_fixed_values}
\title{Generates list of fixed_values based on median/modal values in dataset}
\usage{
generate_fixed_values(data, comp_labels, rounded_zeroes, det_limit)
}
\arguments{
\item{data}{Data used for model development.}

\item{comp_labels}{The labels of the compositional columns.}

\item{rounded_zeroes}{Are zeroes rounded zeroes?}

\item{det_limit}{Detection limit if zeroes are to be imputed. This must be set if \code{rounded_zeroes} is \code{TRUE} and should be the
minimum measurable value in the compositional columns of data.}
}
\value{
dataframe with a single row of fixed_values.
}
\description{
If fixed values are not set, this will set them to modal/ median values.
}
