% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_plotting.R
\name{generate_fixed_values}
\alias{generate_fixed_values}
\title{Generates list of fixed_values based on median/modal values in dataset}
\usage{
generate_fixed_values(data, comp_labels)
}
\arguments{
\item{data}{Data used for model development.}

\item{comp_labels}{The labels of the compositional columns.}
}
\value{
dataframe with a single row of fixed_values.
}
\description{
If fixed values (for non-compositional variables) are not set, this will set them to modal/ median values.
}
