% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilr.R
\name{ilr_trans}
\alias{ilr_trans}
\title{ilr_trans: Ilr transformations}
\usage{
ilr_trans(data)
}
\arguments{
\item{data}{Dataset to have the transformation applied to.}
}
\value{
ilr-transformed columns.
}
\description{
Takes activity columns and returns them after ilr transformation.
}
\examples{
TBC
}
