% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilr.R
\name{ilr_trans}
\alias{ilr_trans}
\title{ilr_trans: Performs ilr transformations (pivot coordinates)}
\usage{
ilr_trans(data)
}
\arguments{
\item{data}{Dataset to have the transformation applied to.}
}
\value{
ilr-transformed columns.
}
\description{
Takes compositional columns and returns them after ilr transformation.
}
\details{
This doesn't use compositions::ilr to insulate it from changes in the choice of
basis used. As pivot coordinates are currently standard in physical activity epidemiology this
is specified.
}
