% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_plotting.R
\name{make_new_data}
\alias{make_new_data}
\title{make_new_data: Generates a new dataset varying in the dimension of interest.}
\usage{
make_new_data(
  from_part,
  to_part,
  fixed_values,
  dataset,
  units = NULL,
  specified_units = NULL,
  comp_labels,
  lower_quantile = 0.05,
  upper_quantile = 0.95,
  granularity = 10000
)
}
\arguments{
\item{from_part}{Should be an element of \code{comp_labels}.}

\item{to_part}{Should be an element of \code{comp_labels}. Should have compositional mean less than \code{from_part}.}

\item{dataset}{Should be dataset used to develop \code{model}.}

\item{units}{What are the units of the compositional variables? Currently available are "unitless" (where working in terms of proportions), "hr/day", "hr/wk", "min/day", "min/wk" and "specified", in which case the \code{specified_units} argument should be set.}

\item{specified_units}{If units are being specified via the composition sum, this is where it is done. It should be a vector where the first argument is a string describing the units, and the second argument is the expected sum of a composition e.g. \code{c("hr/day", 24)}}

\item{comp_labels}{The labels of the compositional columns.}

\item{lower_quantile}{See \code{vary_time_of_interest} and \code{make_new_data}}

\item{upper_quantile}{See \code{vary_time_of_interest} and \code{make_new_data}}

\item{granularity}{Doesn't usually need setting. Parameter indicating how many predictions to make. If too low, plotted curve has gaps. If too high, calculation is slow.}
}
\description{
Generates a new dataset to feed into the plotting functions.
}
