% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_plotting.R
\name{make_new_data}
\alias{make_new_data}
\title{make_new_data: Generates a new dataset varying in the dimension of interest.}
\usage{
make_new_data(
  from_part,
  to_part,
  fixed_values,
  dataset,
  units = NULL,
  specified_units = NULL,
  comp_labels,
  lower_quantile = 0.05,
  upper_quantile = 0.95,
  granularity = 10000
)
}
\arguments{
\item{from_part}{Should be an element of \code{comp_labels}.}

\item{to_part}{Should be an element of \code{comp_labels}.}

\item{fixed_values}{If \code{terms = FALSE}, this gives the fixed values of the non-compositional covariates at which to calculate the prediction. It is generated automatically if not set.}

\item{dataset}{Dataset used to develop \code{model}.}

\item{units}{What should the units of the compositional variables be in any output? Currently available are "unitless" (where working in terms of proportions), "hr/day", "hr/wk", "min/day", "min/wk" and "specified", in which case the \code{specified_units} argument should be set. Note that this doesn't specify the input units, as this is not relevant for any function.}

\item{specified_units}{If units are being specified via the composition sum, this is where it is done. It should be a vector where the first argument is a string describing the units, and the second argument is the expected sum of a composition e.g. \code{c("hr/day", 24)}}

\item{comp_labels}{The labels of the compositional columns.}

\item{lower_quantile}{If set, this gives the lower limit of plotting (as a quantile for both variables of interest).}

\item{upper_quantile}{If set, this gives the upper limit of plotting (as a quantile for both variables of interest).}

\item{granularity}{Does not usually require setting. If set, gives the number of points plotted on the graph. If it is too low, the plot will contain gaps. If it is too high, plotting will be slow.}
}
\description{
Generates a new dataset to feed into the plotting functions.
}
