% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_transfers.R
\name{plot_transfers}
\alias{plot_transfers}
\title{plot_transfers: Plots model predictions.}
\usage{
plot_transfers(from_component, to_component, model, dataset,
  fixed_values = NULL, transformation_type = NULL,
  comparison_component = NULL, component_1 = NULL, comp_labels,
  yllimit = NULL, yulimit = NULL, y_label, plot_log = FALSE,
  lower_quantile = 0.05, upper_quantile = 0.95, units)
}
\arguments{
\item{from_component}{Should be an element of \code{comp_labels}.}

\item{to_component}{Should be an element of \code{comp_labels}. Should have compositional mean less than \code{from_component}.}

\item{dataset}{Should be dataset used to develop \code{model}. Used to set reasonable values to display predictions for based on range of the data.}

\item{fixed_values}{If desired, fixed_values for variables in \code{dataset} which aren't in \code{comp_labels}. These will be used when making predictions}

\item{transformation_type}{Should match transformation used in \code{activity_trans} when developing models.}

\item{comparison_component}{If used, should match transformation used in \code{activity_trans} when developing models.}

\item{component_1}{If used, should match transformation used in \code{activity_trans} when developing models.}

\item{yllimit}{Upper limit to show on y-axis on plot.}

\item{yulimit}{Lower limit to show on y-axis on plot.}

\item{y_label}{Label for y-axis on plot.}

\item{plot_log}{If this is \code{TRUE}, the y-axis will be log-transformed.}

\item{lower_quantile}{See \code{vary_time_of_interest} and \code{make_new_data}}

\item{upper_quantile}{See \code{vary_time_of_interest} and \code{make_new_data}}

\item{units}{What are the units of the compositional variables? E.g. for activity data "hr/day". Currently all non-activity exposure variables should be specified as unitless until support for alternatives units is added.}
}
\value{
Plot with balance of two components plotted as exposure/ independent variable.
}
\description{
Plots model predictions for the transfer of time given.
}
