% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_comp.R
\name{predict_comp}
\alias{predict_comp}
\title{Give model predictions for a particular compositions}
\usage{
predict_comp(data, model, comp_labels, transformation_type = "ilr",
  comparison_component = NULL, component_1 = NULL)
}
\arguments{
\item{comp_labels}{The labels of the compositional columns.}

\item{comparison_component}{Only needed for alr transformation. Should be an element of \code{comp_labels}. Name of component that all other components will be compared to.}

\item{component_1}{Only used if First component in ilr-pivot coordinate transformation. Passed to \code{alter_order_comp_labels} unless \code{NULL}.}

\item{composition_list}{List of compositions}
}
\description{
Give model predictions for a particular compositions
}
