% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_axis_label}
\alias{process_axis_label}
\title{Process axis labels}
\usage{
process_axis_label(label, terms, type)
}
\arguments{
\item{label}{This is the axis label that already exists. To suppress the label entirely, set \code{label == "suppressed"}.}

\item{terms}{Are estimates for differences in outcome associated with differences in compositional variables? If \code{terms = TRUE} all estimates and plots will be for difference in outcome associated with differences in the compositional variables. If \code{terms = FALSE}, \code{fixed_values} is used to set the values of the non-compositional covariates, and outputs are predictions for the outcome based on these values of the non-compositional covariates and the given value of the compositional variables (and confidence intervals include uncertainty due to all variables in the model, not just the compositional variables). Note that for logistic regression models with \code{terms = TRUE} estimates are odds ratios; for logistic regression models with \code{terms = FALSE} estimates are probabilities (i.e. predictions on the response scale).}

\item{type}{Output from \code{process_model_type}}
}
\description{
Process axis labels
}
