% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_units}
\alias{process_units}
\title{Process units argument}
\usage{
process_units(units, specified_units)
}
\arguments{
\item{units}{What are the units of the compositional variables? Currently available are "unitless" (where working in terms of proportions), "hr/day", "hr/wk", "min/day", "min/wk" and "specified", in which case the \code{specified_units} argument should be set.}

\item{specified_units}{If units are being specified via the composition sum, this is where it is done. It should be a vector where the first argument is a string describing the units, and the second argument is the expected sum of a composition e.g. \code{c("hr/day", 24)}}
}
\description{
Process units argument
}
