% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_zeroes}
\alias{process_zeroes}
\title{Process zeroes argument}
\usage{
process_zeroes(data, comp_labels, rounded_zeroes, det_limit = NULL)
}
\arguments{
\item{data}{Dataset to have zeroes imputed for.}

\item{comp_labels}{The labels of the compositional columns.}

\item{rounded_zeroes}{Are zeroes rounded zeroes?}

\item{det_limit}{Detection limit if zeroes are to be imputed. This is needed when \code{rounded_zeroes} is \code{TRUE}. It should be the
minimum measurable value in the compositional columns of data, and should be on the same scale as the (input) compositional columns. If it is
not set, the minimum measured value greater than zero is used.}
}
\description{
Process zeroes argument
}
