% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rescale_det_limit}
\alias{rescale_det_limit}
\title{Rescale det_limit}
\usage{
rescale_det_limit(data, comp_labels, det_limit)
}
\arguments{
\item{data}{Input data.}
}
\description{
Rescale a det_limit given on the same scale as the compositional columns.
}
