% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_of_coefficients.R
\name{tab_covariate_coefs}
\alias{tab_covariate_coefs}
\title{Tabulate coefficients associated with covariates}
\usage{
tab_covariate_coefs(
  model = NULL,
  scale_type = "lp",
  level = 0.95,
  comp_labels,
  transformation_type = "ilr",
  comparison_part = NULL,
  part_1 = NULL
)
}
\arguments{
\item{model}{Model for which extracting coefficients.}

\item{scale_type}{Can take value "lp" (linear predictors) or "exp" (exponentiated- relevant for interpretation of logistic and Cox models).}

\item{level}{The level argument of the confidence intervals. Passed directly to `stats::confint`.}

\item{comp_labels}{The labels of the compositional columns.}

\item{transformation_type}{The type of transformation desired. Should be \code{"ilr", "alr"} or \code{"clr"}.}

\item{comparison_part}{Only needed for alr transformation. Should be an element of \code{comp_labels}. Name of part that all other parts will be compared to.}

\item{part_1}{Only used if first part in ilr-pivot coordinate transformation. Passed to \code{alter_order_comp_labels} unless \code{NULL}.}
}
\value{
Table of covariates and their associated coefficients (parameter value with lower and upper confidence interval of the 95% ).
}
\description{
This pulls out a table of the coefficients of the covariates only in the model.
}
