% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transf_labels}
\alias{transf_labels}
\title{Create columns headings for after the transformation.}
\usage{
transf_labels(comp_labels)
}
\arguments{
\item{comp_labels}{List of compositional column labels.}
}
\description{
Create columns headings for after the transformation.
}
