% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf_variables.R
\name{transf_labels}
\alias{transf_labels}
\title{Create columns headings for after the transformation.}
\usage{
transf_labels(comp_labels, transformation_type,
  comparison_component = NULL, component_1 = NULL)
}
\arguments{
\item{comp_labels}{List of compositional column labels.}
}
\description{
Create columns headings for after the transformation.
}
