% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_comp.R
\name{transform_comp}
\alias{transform_comp}
\title{Transform compositional variables to ilr pivot coordinates}
\usage{
transform_comp(data, comp_labels, rounded_zeroes = TRUE,
  det_limit = NULL, component_1 = NULL)
}
\arguments{
\item{data}{Dataset to have the transformation applied to.}

\item{comp_labels}{The labels of the compositional columns.}

\item{rounded_zeroes}{Are zeroes rounded zeroes?}

\item{det_limit}{Detection limit if zeroes are to be imputed. This must be set if \code{rounded_zeroes} is \code{TRUE} and should be the
minimum measurable value in the compositional columns of data.}

\item{component_1}{First component in ilr-pivot coordinate transformation. Passed to \code{alter_order_comp_labels} unless \code{NULL}.}
}
\value{
\code{data} with ilr-transformed activity columns.
}
\description{
Takes a dataset with a set of labelled compositional columns and returns a dataset with
transformed variables added.
}
\examples{
TBA
}
