% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_comp.R
\name{transform_comp}
\alias{transform_comp}
\title{Transform compositional variables to ilr (pivot), alr or clr coordinates}
\usage{
transform_comp(
  data,
  comp_labels,
  transformation_type = "ilr",
  rounded_zeroes = TRUE,
  det_limit = NULL,
  comparison_part = NULL,
  part_1 = NULL
)
}
\arguments{
\item{data}{Dataset to have the transformation applied to.}

\item{comp_labels}{The labels of the compositional columns.}

\item{transformation_type}{The type of transformation desired. Should be \code{"ilr", "alr"} or \code{"clr"}.}

\item{rounded_zeroes}{Are zeroes rounded zeroes?}

\item{det_limit}{Detection limit if zeroes are to be imputed. This must be set if \code{rounded_zeroes} is \code{TRUE} and should be the
minimum measurable value in the compositional columns of data. It should be on the same scale as the (input) compositional columns.}

\item{comparison_part}{Only needed for alr transformation. Should be an element of \code{comp_labels}. Name of part that all other parts will be compared to.}

\item{part_1}{Only used if First part in ilr-pivot coordinate transformation. Passed to \code{alter_order_comp_labels} unless \code{NULL}.}
}
\value{
\code{data} with transformed activity columns.
}
\description{
Takes a dataset with a set of labelled compositional columns and returns a dataset with
transformed variables added.
}
\examples{
transform_comp(data = simdata,
comp_labels = c("partA", "partB", "partC", "partD", "partE"),
transformation_type = "ilr",
rounded_zeroes = TRUE,
det_limit = 0.0083)

}
