% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_plotting.R
\name{vary_part_of_interest}
\alias{vary_part_of_interest}
\title{Varies part of interest.}
\usage{
vary_part_of_interest(part_of_interest, lower_quantile = 0.05,
  upper_quantile = 0.95, granularity = 10000)
}
\arguments{
\item{part_of_interest}{The variable of interest.}

\item{lower_quantile}{The lower quantile to vary from.}

\item{upper_quantile}{The upper quantile to vary to.
#' @param granularity Doesn't usually need setting. Parameter indicating how many predictions to make. If too low, plotted curve has gaps. If too high, calculation is slow.}
}
\value{
Vector of values going from \code{lower_quantile} to \code{upper_quantile} of the distribution of the varaible of interest.
}
\description{
Produces variable going between percentiles of the part of interest in the data.
}
